requirements.txt
tk==0.1.0
Pillow==9.5.0
numpy==1.24.4
tensorflow==2.2.0